rxlifecycle-lint/src/main/java/com/trello/rxlifecycle/lint/registry/CustomIssueRegistry.java
package com.trello.rxlifecycle.lint.registry;

import java.util.Arrays;
import java.util.List;

import com.android.tools.lint.client.api.IssueRegistry;
import com.android.tools.lint.detector.api.Issue;
import com.trello.rxlifecycle.lint.detectors.RxLifecycleDetector;

/**
 * The list of issues that will be checked when running <code>lint</code>.
 */
@SuppressWarnings("unused")
public class CustomIssueRegistry extends IssueRegistry {

    private List<Issue> mIssues = Arrays.asList(
            RxLifecycleDetector.ISSUE
    );

    public CustomIssueRegistry() {
    }

    @Override
    public List<Issue> getIssues() {
        return mIssues;
    }

}
